Hide matched text, if --license-text is disabled in the scan
