+++ src/pages/ScanInfo/ScanInfo.tsx
@@ -3,42 +3,11 @@ import React, { useEffect, useState } from "react";
 // Maintained Fork of unmaintained but popular react-json-view
 import ReactJson from "@microlink/react-json-view";
 
-import { useWorkbenchDB } from "../../contexts/dbContext";
 import InfoEntry from "./InfoEntry";
+import { useWorkbenchDB } from "../../contexts/dbContext";
+import { ScanInfo, parseIfValidJson, parseScanInfo } from "../../utils/parsers";
 
 import "./scanInfo.css";
-
-interface ScanInfo {
-  tool_name: string;
-  tool_version: string;
-  notice: string;
-  duration: number;
-  options: [string, string][];
-  input: string[];
-  files_count: number;
-  output_format_version: string;
-  spdx_license_list_version: string;
-  operating_system: string;
-  cpu_architecture: string;
-  platform: string;
-  platform_version: string;
-  python_version: string;
-  workbench_version: string;
-  workbench_notice: string;
-  raw_header_content: string;
-}
-
-function parseIfValidJson(str: unknown) {
-  if (typeof str !== "string") return null;
-  try {
-    const parsedObj = JSON.parse(str);
-    // Return only if it is an object & not a primitive value
-    if (Object(parsedObj) === parsedObj) return parsedObj;
-    return null;
-  } catch (e) {
-    return null;
-  }
-}
 
 const ScanInfo = () => {
   const workbenchDB = useWorkbenchDB();
@@ -52,40 +21,7 @@ const ScanInfo = () => {
     db.sync.then(() => {
       db.getScanInfo().then((rawInfo) => {
         console.log("Raw scan info:", rawInfo);
-        const newParsedScanInfo: ScanInfo = {
-          tool_name: rawInfo.getDataValue("tool_name").toString({}) || "",
-          tool_version: rawInfo.getDataValue("tool_version").toString({}) || "",
-          notice: rawInfo.getDataValue("notice").toString({}) || "",
-          duration: Number(rawInfo.getDataValue("duration")),
-          options:
-            Object.entries(
-              parseIfValidJson(rawInfo.getDataValue("options")?.toString({})) ||
-                []
-            ) || [],
-          input:
-            parseIfValidJson(rawInfo.getDataValue("input")?.toString({})) || [],
-          files_count: Number(rawInfo.getDataValue("files_count")),
-          output_format_version:
-            rawInfo.getDataValue("output_format_version")?.toString({}) || "",
-          spdx_license_list_version:
-            rawInfo.getDataValue("spdx_license_list_version")?.toString({}) ||
-            "",
-          operating_system:
-            rawInfo.getDataValue("operating_system")?.toString({}) || "",
-          cpu_architecture:
-            rawInfo.getDataValue("cpu_architecture")?.toString({}) || "",
-          platform: rawInfo.getDataValue("platform")?.toString({}) || "",
-          platform_version:
-            rawInfo.getDataValue("platform_version")?.toString({}) || "",
-          python_version:
-            rawInfo.getDataValue("python_version")?.toString({}) || "",
-          workbench_version:
-            rawInfo.getDataValue("workbench_version")?.toString({}) || "",
-          workbench_notice:
-            rawInfo.getDataValue("workbench_notice")?.toString({}) || "",
-          raw_header_content:
-            rawInfo.getDataValue("header_content")?.toString({}) || "",
-        };
+        const newParsedScanInfo = parseScanInfo(rawInfo);
         console.log("Parsed scan info:", newParsedScanInfo);
         setParsedScanInfo(newParsedScanInfo);
       });
@@ -120,11 +56,11 @@ const ScanInfo = () => {
 
             <InfoEntry
               name="Options"
-              show={parsedScanInfo.options && parsedScanInfo.options.length > 0}
+              show={parsedScanInfo.optionsList && parsedScanInfo.optionsList.length > 0}
             >
               <table className="options-table">
                 <tbody>
-                  {parsedScanInfo.options.map(([key, value]) => (
+                  {parsedScanInfo.optionsList.map(([key, value]) => (
                     <tr key={key}>
                       <td>{key}</td>
                       {typeof value !== "boolean" && <td>{String(value)}</td>}
